+++ app/Http/Controllers/RacketImageController.php
@@ -8,6 +8,7 @@ use App\Models\RacketImage;
 use App\Models\Maker;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Intervention\Image\Facades\Image;
 
 class RacketImageController extends Controller
 {
@@ -45,17 +46,38 @@ class RacketImageController extends Controller
         $validated_request = $request->validated();
 
         try {
-            $file = $request->file('file');
+            // 画像ファイルリサイジング
+            $file = Image::make($request->file('file'));
+            $file->orientate();
+            $file->resize(
+                480,
+                null,
+                function ($constraint) {
+                    // 縦横比を保持したままにする
+                    $constraint->aspectRatio();
+                    // 小さい画像は大きくしない
+                    $constraint->upsize();
+                }
+            );
 
             $filename = now()->format('YmdHis') . $validated_request['title'] . "." . $request->file('file')->extension();
 
-            $path = $file->storeAs('images/rackets', $filename, 'public');
+            // storageに登録するためのpathを生成
+            $storagePath = storage_path('app/public/images/rackets');
+            $fileLocationFullPath = $storagePath . '/' . $filename;
 
-            $racket_image = RacketImage::create([
-                'file_path' => $path,
-                'title' => $validated_request['title'],
-                'maker_id' => $validated_request['maker_id']
-            ]);
+            if ($file->save($fileLocationFullPath)) {
+                // "/var/www/html/strii-backend/storage/app/public/images/rackets/20240105123954リサイズ確認５.jpg"
+                // intervension image導入前の登録の仕様に合わせるため
+                // 上記のようなfullPathをDBのfile_pathカラム用に整形
+                $trimedFilePath = strstr($fileLocationFullPath, 'images');
+
+                $racket_image = RacketImage::create([
+                    'file_path' => $trimedFilePath,
+                    'title' => $validated_request['title'],
+                    'maker_id' => $validated_request['maker_id']
+                ]);
+            }
 
             if (isset($racket_image)) {
                 $maker = Maker::find($racket_image->maker_id);
@@ -69,7 +91,7 @@ class RacketImageController extends Controller
         } catch (\Throwable $e) {
             \Log::error($e);
 
-            return $e;
+            throw $e;
         }
     }
 
@@ -115,14 +137,37 @@ class RacketImageController extends Controller
 
             //新しいファイルがあれば新たにstorageに登録
             if($request->file('file')) {
-                $file = $request->file('file');
+                // 画像ファイルリサイジング
+                $file = Image::make($request->file('file'));
+                $file->orientate();
+                $file->resize(
+                    480,
+                    null,
+                    function ($constraint) {
+                        // 縦横比を保持したままにする
+                        $constraint->aspectRatio();
+                        // 小さい画像は大きくしない
+                        $constraint->upsize();
+                    }
+                );
+
                 $filename = now()->format('YmdHis') . $validated_request['title'] . "." . $request->file('file')->extension();
-                $path = $file->storeAs('images/guts', $filename, 'public');
-    
-                //以前のイメージファイルをstorageフォルダから削除
-                Storage::disk('public')->delete($image->file_path);
-                
-                $image->file_path = $path;
+
+                // storageに登録するためのpathを生成
+                $storagePath = storage_path('app/public/images/rackets');
+                $fileLocationFullPath = $storagePath . '/' . $filename;
+
+                if ($file->save($fileLocationFullPath)) {
+                    // "/var/www/html/strii-backend/storage/app/public/images/rackets/20240105123954リサイズ確認５.jpg"
+                    // intervension image導入前の登録の仕様に合わせるため
+                    // 上記のようなfullPathをDBのfile_pathカラム用に整形
+                    $trimedFilePath = strstr($fileLocationFullPath, 'images');
+
+                    //以前のイメージファイルをstorageフォルダから削除
+                    Storage::disk('public')->delete($image->file_path);
+
+                    $image->file_path = $trimedFilePath;
+                }
             }
 
             $image->title = $validated_request['title'];
@@ -141,7 +186,7 @@ class RacketImageController extends Controller
         } catch (\Throwable $e) {
             \Log::error($e);
 
-            return $e;
+            throw $e;
         }
     }
